Modify: Inventory

+++ src/main/java/com/jannesh/service/InventoryService.java
@@ -1,14 +1,10 @@
 package com.jannesh.service;
 
+import com.jannesh.dto.inventory.InventoryDTO;
 import com.jannesh.dto.inventory.SaveInventoryDTO;
 import com.jannesh.entity.Inventory;
 import com.jannesh.entity.Item;
-import com.jannesh.entity.Vendor;
-import com.jannesh.entity.Warehouse;
 import com.jannesh.repository.InventoryRepository;
-import com.jannesh.repository.ItemRepository;
-import com.jannesh.repository.VendorRepository;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
@@ -17,19 +13,24 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class InventoryService {
     private final InventoryRepository inventoryRepo;
-    private final VendorService vendorService;
-    private final WarehouseService warehouseService;
     private final ItemService itemService;
     private final ModelMapper modelMapper;
 
-    public Inventory createInventory(SaveInventoryDTO requestDTO) {
-        Item item = itemService.createItem(requestDTO.getItem());
+    public InventoryDTO createInventoryDTO(SaveInventoryDTO requestDTO) {
 
-        Inventory inventory = new Inventory();
+        if(inventoryRepo.existsByItem_ItemId(requestDTO.getItemId()))
+            throw new RuntimeException(("Item already exists in the Inventory"));
+
+        Item item = itemService.fetchItemByItemId(requestDTO.getItemId());
+        Inventory inventory = modelMapper.map(requestDTO, Inventory.class);
+
         inventory.setItem(item);
-
-        //Adding Quantity in the Inventory
         inventory.setAvailableQty(requestDTO.getAvailableQty());
+
+        return modelMapper.map(createInventory(inventory), InventoryDTO.class);
+    }
+
+    public Inventory createInventory(Inventory inventory) {
         return inventoryRepo.save(inventory);
     }
